Add USB support for DISCO_L072CZ_LRWAN1
The source code of this commit comes from a work from Jerome Coutant.
The only deviation from this work is that the number of endpoint is set to 8.
In addition, in target.json, USBDEVICE is only added to DISCO_L072CZ_LRWAN1 while
there is other board that should now work with this commit.